# PERF/TYP: typing cast in __getitem__ gives quite some overhead