--- a/pandas/core/arrays/datetimelike.py
+++ b/pandas/core/arrays/datetimelike.py
@@ -331,10 +331,12 @@ class DatetimeLikeArrayMixin(OpsMixin, NDArrayBackedExtensionArray):
         This getitem defers to the underlying array, which by-definition can
         only handle list-likes, slices, and integer scalars
         """
-        # Use cast as we know we will get back a DatetimeLikeArray or DTScalar
-        # result = cast(
-        #     Union[DatetimeLikeArrayT, DTScalarOrNaT], super().__getitem__(key)
-        # )
+        # Use cast as we know we will get back a DatetimeLikeArray or DTScalar,
+        # but skip evaluating the Union at runtime for performance
+        # (see https://github.com/pandas-dev/pandas/pull/44624)
+        result = cast(
+            "Union[DatetimeLikeArrayT, DTScalarOrNaT]", super().__getitem__(key)
+        )
         result = super().__getitem__(key)
         if lib.is_scalar(result):
             return result