While doing some unrelated profiling, I noticed typing overhead in `DatetimeArray.__getitem__`. Isolating that here, there is a typing `cast(..)` call in that method, and removing that as done in the PR, I see the following impact:

```python
arr = pd.date_range("2012", periods=3).array

In [4]: %timeit -r 20 arr[0]
4.76 µs ± 36 ns per loop (mean ± std. dev. of 20 runs, 100000 loops each)  # <-- master
3.44 µs ± 34.4 ns per loop (mean ± std. dev. of 20 runs, 100000 loops each)  # <-- PR
```

(I did this timing several times back and forth on a computer with no other activity at that moment, and the difference was stable)

Personally, I wouldn't have expected such a relative big impact of a typing cast. Because if that's the case, we should be very careful with using that in "hot paths". 

cc @jbrockmendel @simonjayhawkins @Dr-Irv