@jbrockmendel 2021-11-26
wow.  ive been assuming that the casts are negligible.  if that wrong, then there are probably a lot of small improvements to be made by avoiding them across the codebase.

On the other hand, most of this is in the Union:

```
In [1]: from pandas._typing import *

In [2]: from pandas.core.arrays.datetimelike import *

In [3]: %timeit Union[DatetimeLikeArrayT, DTScalarOrNaT]
879 ns ± 19.5 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [4]:  %timeit cast(Union[DatetimeLikeArrayT, DTScalarOrNaT], NaT)
920 ns ± 8.15 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

So at least in this case we can get most of the benefit by just doing the Union once at the module level