@Dr-Irv 2021-11-26
> So at least in this case we can get most of the benefit by just doing the Union once at the module level

Another (maybe easier) solution is to just put the type in double quotes:
```python
        result = cast(
            "Union[DatetimeLikeArrayT, DTScalarOrNaT]", super().__getitem__(key)
        )
```

I did a grep exercise, and that is what is done elsewhere in the code.  The only other place I could find (via grep) that we need to fix this is in `pandas/core/arrays/categorical.py`,  line 517