@jbrockmendel 2021-11-30
@jorisvandenbossche want to change the annotation to a string and call this a win?