@Dr-Irv 2021-12-03
@jorisvandenbossche can you also make a similar change in `pandas/core/arrays/categorical.py:astype()` so we are consistent in terms of doing casts with `Union`